moving navigation to ui
